Sort of got animations working however it's working off of data path and is quite glitchy

+++ editor/src/com/rafkind/paintown/animator/Player.java
@@ -482,7 +482,7 @@ public final class Player extends CharacterStats
 			}
 		});
 	   speedAnim = (JSpinner) controlEditor.find( "speed" );
-	   speedAnim.setValue(new Integer(300));
+	   speedAnim.setValue(new Integer(180));
 		
 		context.add((JComponent)contextEditor.getRootComponent());
